+++ src/main/java/blackjack/domain/user/Player.java
@@ -1,16 +1,16 @@
-package blackjack.domain;
+package blackjack.domain.user;
 
+import blackjack.domain.Card;
 import java.util.List;
 
 public class Player extends Participant {
 
     private static final int PLAYER_DISTRIBUTE_CARD_THRESHOLD = 21;
 
-    private final String name;
+    private final PlayerName playerName;
 
     public Player(final String name) {
-        validateName(name);
-        this.name = name;
+        this.playerName = new PlayerName(name);
     }
 
     @Override
@@ -24,14 +24,6 @@ public class Player extends Participant {
     }
 
     public String getName() {
-        return name;
-    }
-
-    private void validateName(final String name) {
-        boolean isAllLetter = name.chars().allMatch(Character::isLetter);
-        if (isAllLetter) {
-            return;
-        }
-        throw new IllegalArgumentException("이름은 알파벳 소문자만 입력 가능합니다.");
+        return playerName.getName();
     }
 }
